+++ app/views/api/v0/articles/index.json.jbuilder
@@ -3,7 +3,7 @@ json.array! @articles do |article|
   json.id                 article.id
   json.title              article.title
   json.description        article.description
-  json.cover_image        cloud_cover_url article.main_image
+  json.cover_image        cloud_cover_url(article.main_image)
   json.published_at       article.published_at
   json.tag_list           article.cached_tag_list_array
   json.slug               article.slug
@@ -12,6 +12,7 @@ json.array! @articles do |article|
   json.canonical_url      article.processed_canonical_url
   json.comments_count     article.comments_count
   json.positive_reactions_count article.positive_reactions_count
+  json.published_timestamp article.published_timestamp
 
   json.user do
     json.name             article.user.name
