+++ app/views/api/v0/articles/show.json.jbuilder
@@ -13,6 +13,7 @@ json.url                @article.url
 json.canonical_url      @article.processed_canonical_url
 json.comments_count     @article.comments_count
 json.positive_reactions_count @article.positive_reactions_count
+json.published_timestamp @article.published_timestamp
 
 json.body_html @article.processed_html
 json.ltag_style(@article.liquid_tags_used.map { |ltag| Rails.application.assets["ltags/#{ltag}.css"].to_s.html_safe })
