+++ app/views/articles/_single_story.html.erb
@@ -27,8 +27,10 @@
     </div>
   </a>
   <h4>
-    <a href="/<%= story.cached_user.username %>"><%= story.cached_user.name %>・<%= story.readable_publish_date %><span class="time-ago-indicator-initial-placeholder" data-seconds="<%= story.published_at_int %>"></span></a>
-
+    <a href="/<%= story.cached_user.username %>">
+      <%= story.cached_user.name %>・<time datetime="<%= story.published_timestamp %>"><%= story.readable_publish_date %></time>
+      <span class="time-ago-indicator-initial-placeholder" data-seconds="<%= story.published_at_int %>"></span>
+    </a>
   </h4>
   <div class="tags">
     <% story.cached_tag_list_array.each do |tag| %>
