+++ app/views/stories/_main_stories_feed.html.erb
@@ -14,7 +14,8 @@
                      reading_time video_thumbnail_url video video_duration_in_minutes language
                      experience_level_rating experience_level_rating_distribution cached_user cached_organization],
             methods: %i[readable_publish_date cached_tag_list_array flare_tag class_name
-                        cloudinary_video_url video_duration_in_minutes published_at_int],
+                        cloudinary_video_url video_duration_in_minutes published_at_int
+                        published_timestamp],
           ) %>">
     </div>
     <div id="home-articles-object" data-articles="
@@ -23,7 +24,8 @@
                      reading_time video_thumbnail_url video video_duration_in_minutes language
                      experience_level_rating experience_level_rating_distribution cached_user cached_organization],
             methods: %i[readable_publish_date cached_tag_list_array flare_tag class_name
-                        cloudinary_video_url video_duration_in_minutes published_at_int],
+                        cloudinary_video_url video_duration_in_minutes published_at_int
+                        published_timestamp],
           ) %>">
       <% 3.times do %>
         <div class="single-article single-article-small-pic">
