+++ spec/models/article_spec.rb
@@ -507,5 +507,31 @@ RSpec.describe Article, type: :model do
     end
   end
 
+  describe "published_timestamp" do
+    it "returns empty string if the article is new" do
+      expect(Article.new.published_timestamp).to eq("")
+    end
+
+    it "returns empty string if the article is not published" do
+      article.update_column(:published, false)
+      expect(article.published_timestamp).to eq("")
+    end
+
+    it "returns the timestamp of the crossposting date over the publishing date" do
+      crossposted_at = 1.week.ago
+      published_at = 1.day.ago
+      article.update_columns(
+        published: true, crossposted_at: crossposted_at, published_at: published_at,
+      )
+      expect(article.published_timestamp).to eq(crossposted_at.utc.iso8601)
+    end
+
+    it "returns the timestamp of the publishing date if there is no crossposting date" do
+      published_at = 1.day.ago
+      article.update_columns(published: true, crossposted_at: nil, published_at: published_at)
+      expect(article.published_timestamp).to eq(published_at.utc.iso8601)
+    end
+  end
+
   include_examples "#sync_reactions_count", :article
 end
